pyproject.toml
[tool.poetry]
name = "sigrun"
version = "0.1.0"
description = "Sigrun is a Discord bot that manages your Valheim server."
authors = ["Cody"]
readme = "README.md"
packages = [{include = "sigrun"}]

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5"
click = "^7.1.2"
aws-cdk-lib = "^2.49.0"
# "aws-cdk.core" = "^1.139.0"
# "aws-cdk.aws-ec2" = "^1.139.0"
# "aws-cdk.aws-ecs" = "^1.139.0"
# "aws-cdk.aws-ecs-patterns" = "^1.139.0"
# "aws-cdk.aws-apigatewayv2" = "^1.139.0"
# "aws-cdk.aws-apigatewayv2-integrations" = "^1.139.0"
# "aws-cdk.aws-dynamodb" = "^1.139.0"
PyNaCl = "^1.4.0"
httpx = "^0.21.1"
boto3 = "^1.20.24"
rich = "^11.0.0"
"aws-cdk.aws-apigatewayv2-integrations-alpha" = "^2.49.0-alpha.0"
"aws-cdk.aws-apigatewayv2-alpha" = "^2.49.0-alpha.0"
aws-lambda-typing = "^2.16.1"

[tool.poetry.scripts]
sigrun = 'sigrun.cli:sigrun'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
